+++ src/view/_loader.py
@@ -1,12 +1,14 @@
 from __future__ import annotations
 
 import os
-import runpy
+import sys
 import warnings
 from dataclasses import _MISSING_TYPE, Field, dataclass
 from pathlib import Path
 from typing import (TYPE_CHECKING, ForwardRef, Iterable, NamedTuple, TypedDict,
                     get_args, get_type_hints)
+
+from ._util import run_path
 
 try:
     from pydantic.fields import ModelField
@@ -385,7 +387,7 @@ def finalize(routes: list[Route], app: ViewApp):
         target = targets[route.method]
 
         if (not route.path) and (not route.parts):
-            raise InvalidRouteError("route did not specify a path")
+            raise InvalidRouteError(f"{route} did not specify a path")
         lst = virtual_routes.get(route.path or "")
 
         if lst:
@@ -446,14 +448,17 @@ def load_fs(app: ViewApp, target_dir: Path) -> None:
 
     routes: list[Route] = []
 
+    if not target_dir.exists():
+        raise FileNotFoundError(f"{target_dir.absolute()} does not exist")
+
+    sys.path.append(str(target_dir.absolute()))
     for root, _, files in os.walk(target_dir):
         for f in files:
             if f.startswith("_"):
                 continue
 
             path = os.path.join(root, f)
-            Internal.info(f"loading: {path}")
-            mod = runpy.run_path(path)
+            mod = run_path(path)
             current_routes: list[Route] = []
 
             for i in mod.values():
@@ -509,14 +514,18 @@ def load_simple(app: ViewApp, target_dir: Path) -> None:
     Internal.info("loading using simple strategy")
     routes: list[Route] = []
 
+    if not target_dir.exists():
+        raise FileNotFoundError(f"{target_dir.absolute()} does not exist")
+
+    sys.path.append(str(target_dir.absolute()))
+
     for root, _, files in os.walk(target_dir):
         for f in files:
             if f.startswith("_"):
                 continue
 
             path = os.path.join(root, f)
-            Internal.info(f"loading: {path}")
-            mod = runpy.run_path(path)
+            mod = run_path(path)
             mini_routes: list[Route] = []
 
             for i in mod.values():
@@ -534,12 +543,21 @@ def load_simple(app: ViewApp, target_dir: Path) -> None:
 
     finalize(routes, app)
 
+
 def load_patterns(app: ViewApp, target_path: Path) -> None:
     Internal.info("loading using patterns strategy")
-    mod = runpy.run_path(str(target_path))
-    patterns = mod.get("PATTERNS") or mod.get("URL_PATTERNS") or mod.get("URLPATTERNS") or mod.get("urlpatterns") or mod.get("patterns")
+    mod = run_path(str(target_path))
+    patterns = (
+        mod.get("PATTERNS")
+        or mod.get("URL_PATTERNS")
+        or mod.get("URLPATTERNS")
+        or mod.get("urlpatterns")
+        or mod.get("patterns")
+    )
 
     if not patterns:
-        raise RuntimeError
+        raise InvalidRouteError(
+            f"{target_path} did not define a PATTERNS, URL_PATTERNS, URLPATTERNS, urlpatterns, or patterns variable"
+        )
 
     finalize(patterns, app)