+++ src/view/config.py
@@ -6,7 +6,7 @@ from ipaddress import IPv4Address
 from pathlib import Path
 from typing import Any, Dict, Literal, Union
 
-from configzen import ConfigField, ConfigModel
+from configzen import ConfigField, ConfigModel, field_validator
 
 from .exceptions import ViewInternalError
 from .logging import FileWriteMethod, Urgency
@@ -17,6 +17,23 @@ class AppConfig(ConfigModel, env_prefix="view_app_"):
     app_path: str = ConfigField("app.py:app")
     uvloop: Union[Literal["decide"], bool] = "decide"
     loader_path: Path = Path("./routes")
+
+    @field_validator("loader")
+    @classmethod
+    def validate_loader(cls, loader: str):
+        return loader
+
+    @field_validator("loader_path")
+    @classmethod
+    def validate_loader_path(cls, loader_path: Path, values: dict):
+        loader = values["loader"]
+        if loader == "manual":
+            return loader_path
+
+        if (loader == "patterns") and (loader_path == Path("./routes")):
+            return Path("./urls.py").resolve()
+
+        return loader_path.resolve()
 
 
 class ServerConfig(ConfigModel, env_prefix="view_server_"):