+++ tests/fs_routing/delete.py
@@ -1,0 +1,6 @@
+from view import delete
+
+
+@delete()
+async def p():
+    return "delete"