+++ tests/fs_routing/get.py
@@ -1,0 +1,6 @@
+from view import get
+
+
+@get()
+async def g():
+    return "get"