+++ tests/simple_routes/second.py
@@ -1,0 +1,11 @@
+from view import patch, put
+
+
+@put("/put")
+async def pu():
+    return "put"
+
+
+@patch("/patch")
+async def pa():
+    return "patch"