+++ CHANGELOG.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed `EnvironmentError` to `BadEnvironmentError`
 - Added logging functions to `App`
 - Changed environment prefixes for configuration
+- Added `patterns` loader
+- Added handling of relative paths in the configuration setting `loader_path`
+- Added exists validation to `loader_path`
+- Add path to `PATH` environment variable during loading
 
 ## [1.0.0-alpha5] - 2023-09-24